Navigate to selected article almost done. To show info left.

+++ src/main/java/com/mls/Blog_API/service/ArticleService.java
@@ -24,7 +24,7 @@ public class ArticleService implements IArticleService{
     }
 
     @Override
-    public Article searchArticleById(Integer articleId) {
+    public Article searchArticleById(Long articleId) {
         return articleRepository.findById(Long.valueOf(articleId)).orElse(null);
     }
